+++ services/core/java/core-flow-service/src/main/java/org/ubiquia/core/flow/controller/FlowEventController.java
@@ -1,8 +1,12 @@
 package org.ubiquia.core.flow.controller;
 
+import jakarta.transaction.Transactional;
+import jakarta.validation.Valid;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.ubiquia.common.library.api.interfaces.InterfaceEntityToDtoMapper;
@@ -11,6 +15,7 @@ import org.ubiquia.common.library.dao.controller.GenericUbiquiaDaoController;
 import org.ubiquia.common.library.implementation.service.mapper.FlowEventDtoMapper;
 import org.ubiquia.common.model.ubiquia.dto.FlowEvent;
 import org.ubiquia.common.model.ubiquia.entity.FlowEventEntity;
+import org.ubiquia.core.flow.service.registrar.FlowEventRegistrar;
 
 @RestController
 @RequestMapping("/ubiquia/core/flow-service/flow-event")
@@ -23,6 +28,9 @@ public class FlowEventController extends GenericUbiquiaDaoController<FlowEventEn
 
     @Autowired
     private EntityDao<FlowEventEntity> entityDao;
+
+    @Autowired
+    private FlowEventRegistrar flowEventRegistrar;
 
     @Override
     public Logger getLogger() {
@@ -37,5 +45,13 @@ public class FlowEventController extends GenericUbiquiaDaoController<FlowEventEn
     @Override
     public InterfaceEntityToDtoMapper<FlowEventEntity, FlowEvent> getDataTransferObjectMapper() {
         return this.dtoMapper;
+    }
+
+    @PostMapping("/register/post")
+    @Transactional
+    public void register(@RequestBody @Valid final FlowEvent flowEvent) throws Exception {
+        this.getLogger().info("Received FlowEvent registration request for id {}.",
+            flowEvent.getId());
+        this.flowEventRegistrar.tryRegister(flowEvent);
     }
 }